// server/src/types/chatTypes.ts
export type UUID = string;
// ...
export interface MessageEntity {
  messageId: UUID;
  senderId: UUID;
  timestamp: Date;
  status?: MessageDeliveryStatus;
  isEdited?: boolean;
  body: string;
  attachments?: Attachment[];
  reactions?: Reaction[];
}

export type MessageDeliveryStatus = 'Sending' | 'Sent' | 'Delivered' | 'Read' | 'Failed';

export interface Attachment {
  url: string;
  mimeType: string;
  fileName?: string;
  sizeBytes?: number;
  thumbnailUrl?: string;
}

export interface Reaction {
  userId: UUID;
  emoji: string;
}